% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_data.R
\docType{data}
\name{color_data}
\alias{color_data}
\title{Palette data for the kani package}
\format{A \code{list}.}
\usage{
color_data
}
\description{
Data used by the palettes in the kani package
}
\keyword{datasets}
