% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_colour_kani}
\alias{scale_colour_kani}
\title{fivethirtyeight.com color scales}
\usage{
scale_colour_kani(type, ...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Color scales using the colors in the fivethirtyeight graphics.
}
\seealso{
\code{\link{theme_fivethirtyeight}} for examples.

Other colour kani: \code{\link{kani_pal}}
}
