% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_kani}
\alias{theme_kani}
\title{Theme inspired by fivethirtyeight.com plots + my color scheme}
\usage{
theme_kani(base_size = 12, base_family = "Roboto")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme inspired by the plots on
\href{fivethirtyeight.com}{http://fivethirtyeight.com}.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "This is a title",
         subtitle = "This is a subtitle") +
    scale_color_kani() +
    theme_kani()
}

}
